+++ src/WPBT/WPBT_Extras.php
@@ -190,11 +190,6 @@ class WPBT_Extras {
 			static function ( $send, $type ) {
 				$send = 'success' === $type ? false : $send;
 
-				// Disable sending debug email.
-				if ( ! $send ) {
-					add_filter( 'automatic_updates_send_debug_email', '__return_false', 10, 2 );
-				}
-
 				return $send;
 			},
 			10,